+++ classes/Generator/View.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') OR die('No direct script access.');
+/**
+ * Transparent extension of Generator_Generator_View.
+ *
+ * @package    Generator
+ * @category   Views
+ * @author     Zeebee
+ * @copyright  (c) 2012 Zeebee
+ * @license    BSD revised
+ */
+class Generator_View extends Generator_Generator_View {}